fix: show proper error message if secret store is sealed
